% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeHTE.R
\name{ee1.ml}
\alias{ee1.ml}
\title{ee1.ml}
\usage{
ee1.ml(par, dat, par0)
}
\arguments{
\item{par}{parameter.}

\item{dat}{data list.}

\item{par0}{required parameters.}
}
\value{
estimating function values.
}
\description{
An estimating function for the HTE with a continuous outcome.
}
