% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeHTE.R
\name{ee2.ml}
\alias{ee2.ml}
\title{ee2.ml}
\usage{
ee2.ml(par, dat, par0)
}
\arguments{
\item{par}{parameter}

\item{dat}{data list}

\item{par0}{required parameters}
}
\value{
estimating function values
}
\description{
An estimating function for the HTE with a binary outcome.
}
