% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeHTE.R
\name{estforboot}
\alias{estforboot}
\title{estforboot}
\usage{
estforboot(A.t, X.t, Y.t, A.os, X.os, Y.os, n.t, m, cnames, family.oc,
  nptb, gamma1, gamma2)
}
\arguments{
\item{A.t}{is a vector of treatment  ( n x 1).}

\item{X.t}{is a matrix of covariates without intercept  ( n x p).}

\item{Y.t}{is a vector of outcome  ( n x 1).}

\item{A.os}{is a vector of treatment  ( m x 1).}

\item{X.os}{is a matrix of covariates without intercept  ( m x p).}

\item{Y.os}{is a vector of outcome  ( m x 1).}

\item{n.t}{is the RT sample size.}

\item{m}{is the RW sample size.}

\item{cnames}{is the names for the estimators.}

\item{family.oc}{specifies the family for the outcome model.
\code{"gaussian"}: a linear regression model for the continuous outcome.
\code{"binomial"}: a logistic regression model for the binary outcome.}

\item{nptb}{is the number for replication-based method for estimating Sigma_SS.}

\item{gamma1}{is the parameter in the violation test such that T > chisq_{1-gamma1} indicates the bias of the RW analysis is significant.}

\item{gamma2}{is a smooth parameter in using the normal cummulative distribution with mean 0 and variance gamma2^2 function to approximate the indicator function.}
}
\value{
the boostrap statistics
}
\description{
A function computes the bootstrap statistics.
}
